+++ src/Fields/TextAreaField.tsx
@@ -40,6 +40,9 @@ export function TextAreaField<
         value={field.value ?? undefined}
         disabled={disabled}
         {...component}
+        // Avoid losing focus when triggering/resolving a validation error
+        // https://4x.ant.design/components/input/#Why-Input-lose-focus-when-change-prefix/suffix/showCount
+        showCount={component?.showCount || { formatter: () => '' }}
       />
     </FieldWrapper>
   );