+++ src/Fields/index.stories.tsx
@@ -13,6 +13,7 @@ import { InputField } from './InputField';
 import { InputNumberField } from './InputNumberField';
 import { RadioGroupField } from './RadioGroupField';
 import { SelectField } from './SelectField';
+import { SwitchField } from './SwitchField';
 import { TextAreaField } from './TextAreaField';
 import { formItemFieldProps } from './formItemFieldProps';
 
@@ -21,20 +22,6 @@ import { Space } from '../Space';
 
 export default {
   title: 'Fields',
-  argTypes: {
-    disabled: {
-      control: {
-        type: 'boolean',
-        checked: false,
-      },
-    },
-    hasError: {
-      control: {
-        type: 'boolean',
-        checked: false,
-      },
-    },
-  },
 };
 
 type FormType = {
@@ -43,14 +30,29 @@ type FormType = {
   input_number: number;
   radio_group: 1 | 2;
   select: 'a' | 'b';
+  switch: boolean;
   text_area: string;
 };
 
 export const Default: Story<{
-  hasError: boolean;
   disabled: boolean;
 }> = function Default(props) {
-  const formMethods = useForm<FormType>();
+  const formMethods = useForm<FormType>({ mode: 'all' });
+
+  return (
+    <FieldProvider {...props}>
+      <FormProvider {...formMethods}>
+        <AllFields />
+      </FormProvider>
+    </FieldProvider>
+  );
+};
+
+export const ControlledError: Story<{
+  hasError: boolean;
+  disabled: boolean;
+}> = function ControlledError(props) {
+  const formMethods = useForm<FormType>({ mode: 'all' });
 
   const { hasError, ...propsRest } = props;
 
@@ -79,55 +81,45 @@ export const Default: Story<{
     </FieldProvider>
   );
 };
+ControlledError.argTypes = {
+  disabled: {
+    control: {
+      type: 'boolean',
+      checked: false,
+    },
+  },
+  hasError: {
+    control: {
+      type: 'boolean',
+      checked: false,
+    },
+  },
+};
 
-export const NestedProviders: Story<
-  FieldProviderProps & { hasError: boolean }
-> = function NestedProviders(props) {
-  const formMethods = useForm<FormType>();
-
-  const { hasError } = props;
-
-  const parentFormItemProps = () =>
-    formItemFieldProps({
-      invalid: false,
-      error: undefined,
-      isTouched: false,
-      isDirty: false,
+export const NestedProviders: Story<FieldProviderProps> =
+  function NestedProviders(props) {
+    const formMethods = useForm<FormType>({
+      mode: 'all',
     });
 
-  const childFormItemProps = useCallback(
-    () =>
-      formItemFieldProps({
-        invalid: hasError,
-        error: hasError
-          ? {
-              type: 'validate',
-              message: 'some error',
-            }
-          : undefined,
-
-        isTouched: true,
-        isDirty: true,
-      }),
-    [hasError],
-  );
-
-  return (
-    <FieldProvider
-      disabled={!props.disabled}
-      formItemProps={parentFormItemProps}
-    >
-      {/* overwrite upper provider values */}
-      <FieldProvider
-        disabled={props.disabled}
-        formItemProps={childFormItemProps}
-      >
-        <FormProvider {...formMethods}>
-          <AllFields />
-        </FormProvider>
+    return (
+      <FieldProvider disabled={!props.disabled}>
+        {/* overwrite upper provider values */}
+        <FieldProvider disabled={props.disabled}>
+          <FormProvider {...formMethods}>
+            <AllFields />
+          </FormProvider>
+        </FieldProvider>
       </FieldProvider>
-    </FieldProvider>
-  );
+    );
+  };
+NestedProviders.argTypes = {
+  disabled: {
+    control: {
+      type: 'boolean',
+      checked: false,
+    },
+  },
 };
 
 function AllFields() {
@@ -145,6 +137,7 @@ function AllFields() {
       </CheckboxField>
       <InputField
         name="input"
+        rules={{ required: 'You really need this', maxLength: 3 }}
         control={formMethods.control}
         formItem={{
           label: 'Input Label',
@@ -152,6 +145,7 @@ function AllFields() {
       />
       <InputNumberField
         name="input_number"
+        rules={{ required: 'Absolutely necessary' }}
         control={formMethods.control}
         formItem={{
           label: 'InputNumber Label',
@@ -177,6 +171,13 @@ function AllFields() {
           options: ['a', 'b'].map(toFormInputOption),
         }}
       />
+      <SwitchField
+        name="switch"
+        control={formMethods.control}
+        formItem={{
+          label: 'Switch Label',
+        }}
+      />
       <TextAreaStory />
     </Form>
   );
@@ -198,6 +199,7 @@ function TextAreaStory() {
           ref: textArea1Ref,
           minLength: 3,
         }}
+        rules={{ required: 'Very necessary' }}
       />
       <TextAreaField
         name="text_area"