+++ src/ImageMap/index.tsx
@@ -1,0 +1,5 @@
+export { PopoverArea } from './PopoverArea';
+export { HiddenArea } from './HiddenArea';
+
+export { TecanLayout } from './TecanLayout';
+export { ImageMap, ImageMapProps } from './ImageMap';