+++ src/Input/index.stories.tsx
@@ -2,7 +2,12 @@ import { Story } from '@storybook/react';
 import React, { useState } from 'react';
 
 import { NumericIDInput, NumericIDInputProps } from './NumericIDInput';
-import { InputNumberProps, InputProps, TextAreaProps } from './common';
+import {
+  InputNumberProps,
+  InputProps,
+  SearchProps,
+  TextAreaProps,
+} from './common';
 
 import { Input, InputNumber } from './index';
 
@@ -19,8 +24,12 @@ export const TextArea: Story<TextAreaProps> = function TextArea(args) {
   return <Input.TextArea {...args} />;
 };
 
+export const Search: Story<SearchProps> = function Search(args) {
+  return <Input.Search {...args} />;
+};
+
 export const Number: Story<InputNumberProps> = function Number(args) {
-  const [num, setNum] = useState<number>(0);
+  const [num, setNum] = useState<number | null>(0);
 
   return (
     <InputNumber
@@ -28,7 +37,7 @@ export const Number: Story<InputNumberProps> = function Number(args) {
       max={10}
       value={num}
       onChange={(val) =>
-        setNum(typeof val !== 'number' ? parseInt(val, 10) : val)
+        setNum(typeof val === 'string' ? parseInt(val, 10) : val)
       }
       {...args}
     />