+++ src/Popover/index.tsx
@@ -1,0 +1,4 @@
+import { Popover as AntdPopover, PopoverProps as AntdPopoverProps } from 'antd';
+
+export const Popover: typeof AntdPopover = AntdPopover;
+export type PopoverProps = AntdPopoverProps;