+++ src/Table/index.tsx
@@ -12,7 +12,7 @@ export {
   ColumnProps,
   TablePaginationConfig,
 } from 'antd/es/table';
-export { SortOrder } from 'antd/es/table/interface';
+export { FilterDropdownProps, SortOrder } from 'antd/es/table/interface';
 
 export type TableProps<RecordType> = AntdTableProps<RecordType>;